Move process submission model to models
